[core-lro] merge azureAsync and location strategies
